requirements.txt
# HTTP requests and scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# NLP for keyword extraction (legacy)
rake-nltk>=1.0.6
nltk>=3.8.1

# Google Gemini AI
google-genai>=1.0.0

# Environment and configuration
python-dotenv>=1.0.0

# Output formatting
tabulate>=0.9.0
rich>=13.7.0

# CLI interface
click>=8.1.0

# Web interface
flask>=3.0.0

# Rate limiting
ratelimit>=2.2.1
